# docs/submission/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SAHYOG Blockchain Intelligence",
    "description": "Complaint intake, replayable ingestion, tracing, attribution and VASP federation.",
    "version": "0.16.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Login",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/2fa/setup": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Setup",
        "operationId": "setup_api_v1_auth_2fa_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetupInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/2fa/confirm": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Confirm",
        "operationId": "confirm_api_v1_auth_2fa_confirm_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/2fa/verify": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Verify",
        "operationId": "verify_api_v1_auth_2fa_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TotpInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases": {
      "post": {
        "tags": [
          "Cases"
        ],
        "summary": "Create Case",
        "operationId": "create_case_api_v1_cases_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Cases"
        ],
        "summary": "List Cases",
        "operationId": "list_cases_api_v1_cases_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/CaseStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "chain",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Chain"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Chain"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CasePage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/extract-from-image": {
      "post": {
        "tags": [
          "Cases"
        ],
        "summary": "Extract",
        "operationId": "extract_api_v1_cases_extract_from_image_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_extract_api_v1_cases_extract_from_image_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/attachments/{attachment_id}/image": {
      "get": {
        "tags": [
          "Cases"
        ],
        "summary": "Attachment Image",
        "operationId": "attachment_image_api_v1_cases_attachments__attachment_id__image_get",
        "parameters": [
          {
            "name": "attachment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attachment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}": {
      "get": {
        "tags": [
          "Cases"
        ],
        "summary": "Get Case",
        "operationId": "get_case_api_v1_cases__case_id__get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{case_id}/victim-transactions": {
      "post": {
        "tags": [
          "Cases"
        ],
        "summary": "Attach",
        "operationId": "attach_api_v1_cases__case_id__victim_transactions_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttachmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "sha256": {
            "type": "string",
            "title": "Sha256"
          },
          "extracted_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Extracted Json"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "sha256",
          "extracted_json",
          "created_at"
        ],
        "title": "AttachmentOut"
      },
      "Body_extract_api_v1_cases_extract_from_image_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_extract_api_v1_cases_extract_from_image_post"
      },
...
      "CaseCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 160,
            "minLength": 3,
            "title": "Title",
            "default": "Blockchain complaint"
          },
          "sahyog_ref": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9_/-]{3,80}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sahyog Ref"
          },
          "transaction": {
            "$ref": "#/components/schemas/TransactionInput"
          },
          "attachment_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 10,
            "title": "Attachment Ids"
          },
          "extraction_reviewed": {
            "type": "boolean",
            "title": "Extraction Reviewed",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "transaction"
        ],
        "title": "CaseCreate"
      },
      "CaseOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "case_ref": {
            "type": "string",
            "title": "Case Ref"
          },
          "sahyog_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sahyog Ref"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "gang_case_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gang Case Id"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "transactions": {
            "items": {
              "$ref": "#/components/schemas/TransactionOut"
            },
            "type": "array",
            "title": "Transactions"
          },
          "attachments": {
            "items": {
              "$ref": "#/components/schemas/AttachmentOut"
            },
            "type": "array",
            "title": "Attachments"
          }
        },
        "type": "object",
        "required": [
          "id",
          "case_ref",
          "sahyog_ref",
          "title",
          "status",
          "gang_case_id",
          "scenario_id",
          "created_by",
          "created_at",
          "source",
          "transactions",
          "attachments"
        ],
        "title": "CaseOut"
      },
      "CasePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CaseOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "CasePage"
      },
      "CaseStatus": {
        "type": "string",
        "enum": [
          "OPEN",
          "TRACING",
          "ATTRIBUTED",
          "FREEZE_PENDING",
          "CLOSED"
        ],
        "title": "CaseStatus"
      },
      "Chain": {
        "type": "string",
        "enum": [
          "ethereum",
          "bnb",
          "polygon",
          "tron",
          "bitcoin"
        ],
        "title": "Chain"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoginInput": {
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "title": "LoginInput"
      },
...
      "SetupInput": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 72,
            "minLength": 1,
            "title": "Password"
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "SetupInput"
      },
...
      "TotpInput": {
        "properties": {
          "code": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "TotpInput"
      },
...
      "TransactionInput": {
        "properties": {
          "chain": {
            "$ref": "#/components/schemas/Chain"
          },
          "tx_hash": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 66
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Hash"
          },
          "victim_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Victim Address"
          },
          "suspect_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Suspect Address"
          },
          "token": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z0-9._-]{1,20}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{1,38}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "decimals": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 30.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Decimals"
          },
          "tx_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Time"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "chain"
        ],
        "title": "TransactionInput"
      },
      "TransactionOut": {
        "properties": {
          "chain": {
            "$ref": "#/components/schemas/Chain"
          },
          "tx_hash": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 66
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Hash"
          },
          "victim_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Victim Address"
          },
          "suspect_address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Suspect Address"
          },
          "token": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z0-9._-]{1,20}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token"
          },
          "amount": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[0-9]{1,38}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "decimals": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 30.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Decimals"
          },
          "tx_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tx Time"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "chain",
          "id",
          "status"
        ],
        "title": "TransactionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}